backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-docx==0.8.11
docxtpl==0.16.7
sqlalchemy==2.0.23
pydantic==2.4.2
pillow==10.1.0
aiofiles==23.2.1
jinja2==3.1.2
python-jose==3.3.0
passlib==1.7.4
